+++ src/core/automator/environment.ts
@@ -6,7 +6,9 @@ import { player } from '../global';
 import Decimal from 'break_eternity.js';
 import { Hydra } from '../hydra/hydra';
 import { Dilute } from '../hydra/dilute';
-import { importMusic } from './music-play';
+import { importMusic } from './automator-modules/music-play';
+import { importMath } from './automator-modules/math';
+import { importHydra } from './automator-modules/hydra';
 
 export class Environment {
 	parent: Environment | null = null;
@@ -83,25 +85,7 @@ class ToStringFunction extends Callable {
 const putf = new PutFunction();
 const delayf = new DelayFunction();
 const toStringFunction = new ToStringFunction();
-const maxFunction = new (class MaxFunction extends Callable {
-	async call(env: Environment, ...args: any[]) {
-		let max = new Decimal(-Infinity);
-		for (const number of args) {
-			console.log(max, number);
-			max = max.max(number);
-		}
-		return max;
-	}
-})();
-const minFunction = new (class MinFunction extends Callable {
-	async call(env: Environment, ...args: any[]) {
-		let min = new Decimal(Infinity);
-		for (const number of args) {
-			min = min.min(number);
-		}
-		return min;
-	}
-})();
+
 const getFunction = new (class GetFunction extends Callable {
 	async call(env: Environment, ...args: any[]) {
 		if (args[0].get) {
@@ -138,90 +122,13 @@ parentEnvironment.set('set', setFunction);
 parentEnvironment.set('player', getPlayerData);
 parentEnvironment.isReadonly = true;
 
-const hydraResetFunction = new (class hydraReset extends Callable {
-	async call(env: Environment, ...args: any[]) {
-		if (args[0] instanceof Decimal) {
-			if (args[0].eq(0)) {
-				Hydra.hydraReset();
-				return;
-			} else if (args[0].eq(1)) {
-				Hydra.prestige(0);
-				return;
-			} else if (args[0].eq(2)) {
-				Hydra.prestige(1);
-				return;
-			} else if (args[0].eq(3)) {
-				Hydra.prestige(2);
-				return;
-			} else if (args[0].eq(4)) {
-				Hydra.prestige(3);
-				return;
-			}
-		}
-		throw new TypeError('错误的:hydra.reset参数');
-	}
-})();
-const hydraDiluteFunction = new (class hydraReset extends Callable {
-	async call(env: Environment, ...args: any[]) {
-		if (args[0] === undefined) {
-			return Dilute.diluteButton();
-		}
-		if (args[0] instanceof Decimal) {
-		}
-		throw new TypeError('错误的:hydra.dilute参数');
-	}
-})();
-const hydraDiluteSetFunction = new (class hydraReset extends Callable {
-	async call(env: Environment, ...args: any[]) {
-		if (args[0] instanceof Decimal && args[1] instanceof Decimal) {
-			if (player.hydra.dilute.inDilute) return;
-			const i = args[0].toNumber();
-			const v = args[1].toNumber();
-			if ([1, 2, 3, 4, 5, 6, 7, 8, 9].includes(i)) {
-				if (i == 7 || i == 8 || i == 9) {
-					player.hydra.dilute.solvent[i - 1] = !!v;
-				} else {
-					player.hydra.dilute.solvent[i - 1] = v;
-				}
-				return;
-			}
-		}
-		throw new TypeError('错误的:hydra.dilute参数');
-	}
-})();
 export function tryInclude(pkg: string) {
 	if (pkg == 'math') {
-		const readonlyDictionary = new Dictionary();
-		readonlyDictionary.set('max', maxFunction);
-		readonlyDictionary.set('min', minFunction);
-		readonlyDictionary.readonly = true;
-		parentEnvironment.isReadonly = false;
-		parentEnvironment.set('math', readonlyDictionary);
-		parentEnvironment.isReadonly = true;
+		importMath(parentEnvironment);
 		return;
 	}
 	if (pkg == 'hydra' && player.upgrades['ts_auto_pkg_hydra']) {
-		const readonlyDictionaryHydra = new (class extends Dictionary {
-			get(key: any) {
-				if (key !== 'deduces') {
-					return Dictionary.prototype.get.call(this, key);
-				}
-				if (key == 'deduces') {
-					return new AutomatorArray([player.hydra.deduceOrdinal[0].add(1).sub(1)]);
-				}
-				return undefined;
-			}
-			set(key: any, value: any) {
-				return Dictionary.prototype.set.call(this, key, value);
-			}
-		})();
-		readonlyDictionaryHydra.set('reset', hydraResetFunction);
-		readonlyDictionaryHydra.set('dilute', hydraDiluteFunction);
-		readonlyDictionaryHydra.set('diluteset', hydraDiluteSetFunction);
-		readonlyDictionaryHydra.readonly = true;
-		parentEnvironment.isReadonly = false;
-		parentEnvironment.set('hydra', readonlyDictionaryHydra);
-		parentEnvironment.isReadonly = true;
+		importHydra(parentEnvironment);
 		return;
 	}
 	if (pkg == 'music') {
